QNX/BB10: Content browser and navigation fixes
Fix HTTP page navigation on QNX. Bypass DevTools instrumentation,
UKM source ID generation, and crash key serialization that hang in
single-process mode. Fix NavigationRequest::OnResponseStarted to
properly handle cross-origin policy and frame host selection. Add
QNX-specific process launch, sandbox, and zygote stubs.

+++ content/browser/aggregation_service/aggregatable_report_scheduler.cc
@@ -101,7 +101,8 @@ AggregatableReportScheduler::GetFailedReportDelay(int failed_send_attempts) {
     return absl::nullopt;
 
   return kInitialRetryDelay *
-         std::pow(kRetryDelayFactor, failed_send_attempts - 1);
+         std::pow(static_cast<double>(kRetryDelayFactor),
+                  static_cast<double>(failed_send_attempts - 1));
 }
 
 AggregatableReportScheduler::TimerDelegate::TimerDelegate(
